+++ pennylane_sphinx_theme/footer.py
@@ -151,7 +151,7 @@ FOOTER = {
                 },
                 {
                     "name": "Devices",
-                    "href": "https://pennylane.ai/plugins",
+                    "href": "https://pennylane.ai/devices",
                 },
                 {
                     "name": "Catalyst",